+++ core/lib/http-server.ts
@@ -1,18 +1,47 @@
 import { Event } from '../deps.ts';
+import { ContextRegistry } from './context/context-registry.ts';
+import type { Context } from './context/context.ts';
 import type { Controller } from './controller.ts';
 import { HttpError } from './errors/http.error.ts';
 import { MiddlewareChain } from './middleware-chain.ts';
 import type {
-  BaseRequestContext,
-  BeforeRespondHandler,
-  DefaultRequestContext,
-  DefaultResponseHandler,
+  BeforeRespondListener,
+  ContextGetter,
+  DefaultResponseListener,
   HttpMethod,
   Middleware,
   RequestHandler,
-  ServerErrorHandler,
+  ServerErrorListener,
 } from './model.ts';
 import { baseMatchesPath, getRequestPath } from './util.ts';
+
+/**
+ * A builder that creates an `HttpServer` instance.
+ */
+/*
+ * Note: This is included because Deno requires all exposed members to be explicitly typed:
+ * https://jsr.io/docs/about-slow-types. Therefore, the type of the static `Builder` class
+ * can't be inferred, it must be explicitly defined. Because of the hacky way inner classes
+ * work with JS, you can't target the type of an inner class with something like `HttpServer.Builder`
+ * for TS. So, the only way to be able to explcitly define the type of the builder is to
+ * quite literally explicitly define it here and then say that `Builder` is a class that
+ * implements this interface.
+ *
+ * I don't like the fact this means double maintenance every time I want to add something to the
+ * builder's exposed members, but Deno/JSR doesn't really give me a lot of options in this
+ * case.
+ */
+export interface HttpServerBuilder {
+  build(): HttpServer;
+  base(basePath: string): HttpServerBuilder;
+  defaultResponseHandler(handler: (req: Request) => Response): HttpServerBuilder;
+  entryMiddleware(...middleware: Middleware[]): HttpServerBuilder;
+  controller(c: Controller): HttpServerBuilder;
+  ssl(sslOptions: Deno.TlsCertifiedKeyPem): HttpServerBuilder;
+  errorListeners(...listeners: ServerErrorListener[]): HttpServerBuilder;
+  defaultResponseListeners(...listeners: DefaultResponseListener[]): HttpServerBuilder;
+  beforeRespondListeners(...listeners: BeforeRespondListener[]): HttpServerBuilder;
+}
 
 /**
  * Enables configuring and running a lightweight HTTP server. The server is 
@@ -60,28 +89,32 @@ import { baseMatchesPath, getRequestPath } from './util.ts';
  * server encounters an uncaught error that's _not_ an HttpError, it responds with
  * a 500.
  */
-export class HttpServer<RequestContext extends BaseRequestContext = DefaultRequestContext> {
+export class HttpServer {
   /**
    * Triggered when the server encounters an uncaught error. This can occur
    * when middleware or a request handler throws an error that it doesn't catch.
    *
    * The server catches all errors and any error that's not a specific
    * error recognized by the server results in a 500 response.
+   *
+   * Subscribed handlers should not throw.
    */
-  onError: Event<ServerErrorHandler> = new Event<ServerErrorHandler>();
+  onError: Event<ServerErrorListener> = new Event<ServerErrorListener>();
   /**
    * Triggered when the server determines it has no controller to handle the
    * incoming request and is going to send the default response.
    */
-  onDefault: Event<DefaultResponseHandler> = new Event<DefaultResponseHandler>();
+  onDefaultResponse: Event<DefaultResponseListener> = new Event<DefaultResponseListener>();
   /**
    * Triggered when the server is about to send a response.
+   *
+   * Subscribed handlers should not throw.
    */
-  onBeforeRespond: Event<BeforeRespondHandler> = new Event<BeforeRespondHandler>();
+  onBeforeRespond: Event<BeforeRespondListener> = new Event<BeforeRespondListener>();
 
   #base?: string;
-  #controllers: Controller<RequestContext>[] = [];
-  #entryMiddlewareChain: MiddlewareChain<RequestContext> = new MiddlewareChain();
+  #controllers: Controller[] = [];
+  #entryMiddlewareChain: MiddlewareChain = new MiddlewareChain();
   #ssl?: Deno.TlsCertifiedKeyPem;
   #httpServer: Deno.HttpServer | undefined;
   #defaultResponseHandler: (req: Request) => Response = () => new Response(undefined, { status: 404 });
@@ -128,96 +161,7 @@ export class HttpServer<RequestContext extends BaseRequestContext = DefaultReque
     await this.#httpServer?.finished;
   }
 
-  /**
-   * Sets the server's base path. This is the path that a client must include
-   * to talk to the server's controllers. For example, if you set this to `/api`, clients
-   * must start their request paths with `/api`, otherwise the server won't
-   * send the request to one of its controllers.
-   *
-   * If this isn't set, there's no required base path to talk to the server's controllers.
-   *
-   * @param basePath - The base path to use.
-   *
-   * @returns The server instance.
-   */
-  base(basePath: string): HttpServer<RequestContext> {
-    this.#base = basePath;
-
-    return this;
-  }
-
-  /**
-   * Sets the handler that will be called when the server determines
-   * it has no controller that can handle the incoming request.
-   *
-   * If this isn't set, the default handlder returns a 404 response with no body.
-   *
-   * @param handler
-   *
-   * @returns The server instance.
-   */
-  defaultResponseHandler(handler: (req: Request) => Response): HttpServer<RequestContext> {
-    this.#defaultResponseHandler = handler;
-
-    return this;
-  }
-
-  /**
-   * Adds the provided middleware to the entry middleware chain.
-   * Entry middleware runs when a request enters the server. It runs before
-   * anything else, and runs even if the client's request didn't include
-   * the server's `base` if you provided one.
-   *
-   * Middleware runs in the order provided.
-   *
-   * This method is additive and may be called any number of times. Calling
-   * it more than once won't remove the middleware from previous invocations.
-   * Middleware given to subsequent invocations are appended to any middleware
-   * given in previous invocations.
-   * 
-   * @param middleware - The middleware to add.
-   *
-   * @returns The server instance.
-   */
-  entryMiddleware(...middleware: Middleware<RequestContext>[]): HttpServer<RequestContext> {
-    this.#entryMiddlewareChain.add(...middleware);
-
-    return this;
-  }
-
-  /**
-   * Adds the specified controller to the server. Controllers are the
-   * heart of the server and inform it on what requests it can and cannot
-   * handle.
-   *
-   * If a request comes in and the server can't find an appropriate
-   * request handler in one of its registered controllers, it will send
-   * back the default response.
-   *
-   * @param c - The controller to add.
-   *
-   * @returns The server instance.
-   */
-  controller(c: Controller<RequestContext>): HttpServer<RequestContext> {
-    this.#controllers.push(c);
-
-    return this;
-  }
-
-  /**
-   * Allows you to configure the server to handle HTTPS traffic.
-   *
-   * @param sslOptions - Object containing parameters to customize how the server manages TLS.
-   *
-   * @returns The server instance.
-   */
-  ssl(sslOptions: Deno.TlsCertifiedKeyPem): HttpServer<RequestContext> {
-    this.#ssl = sslOptions;
-
-    return this;
-  }
-
-  getHandlingController(path: string): Controller<RequestContext> | undefined {
+  getHandlingController(path: string): Controller | undefined {
     if (baseMatchesPath(this.#base, path)) {
       return this.#controllers.find((controller) => controller.matchesPath(this.getPathWithoutServerBase(path)));
     }
@@ -231,19 +175,23 @@ export class HttpServer<RequestContext extends BaseRequestContext = DefaultReque
 
   #handleHttpRequest: Deno.ServeHandler = async (req, info) => {
     const mutableHeaders = new Headers();
+    const contextRegistry = new ContextRegistry();
 
     let res: Response;
     try {
       const reqPath = getRequestPath(req);
-      // Can't currently figure out how to get the types to play well here.
-      const ctx = {} as unknown as RequestContext;
 
       const entryMiddlewareResult = await this.#entryMiddlewareChain.run({
         req,
         resHeaders: mutableHeaders,
         server: this,
-        remoteAddr: info.remoteAddr,
-        ctx,
+        remoteAddr: info.remoteAddr as Deno.NetAddr,
+        getContext: <T>(context: Context<T>): T => {
+          return contextRegistry.getContext(context);
+        },
+        setContext: <T>(context: Context<T>, value: T): void => {
+          contextRegistry.register(context, value);
+        },
       });
 
       const { controller, handler, params } = this.#getRequestHandlerDetails(req.method as HttpMethod, reqPath) ?? {};
@@ -251,21 +199,33 @@ export class HttpServer<RequestContext extends BaseRequestContext = DefaultReque
       if (entryMiddlewareResult) {
         res = entryMiddlewareResult;
       } else if (controller && handler) {
+        const controllerScopedContextGetter: ContextGetter = <T>(context: Context<T>): T => {
+          return contextRegistry.getContext(context, controller.constructor);
+        };
+
         const controllerMiddlewareResult = await controller.middlewareChain.run({
           req,
           resHeaders: mutableHeaders,
           server: this,
-          remoteAddr: info.remoteAddr,
-          ctx,
+          remoteAddr: info.remoteAddr as Deno.NetAddr,
+          getContext: controllerScopedContextGetter,
+          setContext: <T>(context: Context<T>, value: T): void => {
+            contextRegistry.register(context, value, controller.constructor);
+          },
         });
 
         if (controllerMiddlewareResult) {
           res = controllerMiddlewareResult;
         } else {
-          res = await handler({ req, params: params ?? {}, ctx, remoteAddr: info.remoteAddr });
+          res = await handler({
+            req,
+            params: params ?? {},
+            remoteAddr: info.remoteAddr as Deno.NetAddr,
+            getContext: controllerScopedContextGetter,
+          });
         }
       } else {
-        this.onDefault.trigger(req);
+        this.onDefaultResponse.trigger(req);
 
         res = this.#defaultResponseHandler(req);
       }
@@ -303,8 +263,8 @@ export class HttpServer<RequestContext extends BaseRequestContext = DefaultReque
     path: string
   ):
     | {
-        controller: Controller<RequestContext>;
-        handler?: RequestHandler<RequestContext>;
+        controller: Controller;
+        handler?: RequestHandler;
         params?: Record<string, string>;
       }
     | undefined {
@@ -322,4 +282,164 @@ export class HttpServer<RequestContext extends BaseRequestContext = DefaultReque
 
     return undefined;
   }
+
+  /**
+   * Builder class to aid with configuring and building an `HttpServer` instance.
+   *
+   * **If you're using this for a type annotation**: don't. Use the `HttpServerBuilder` type instead.
+   */
+  static readonly Builder: { new (): HttpServerBuilder } = class Builder implements HttpServerBuilder {
+    #server: HttpServer;
+
+    constructor() {
+      this.#server = new HttpServer();
+    }
+
+    build(): HttpServer {
+      return this.#server;
+    }
+
+    /**
+     * Sets the server's base path. This is the path that a client must include
+     * to talk to the server's controllers. For example, if you set this to `/api`, clients
+     * must start their request paths with `/api`, otherwise the server won't
+     * send the request to one of its controllers.
+     *
+     * If this isn't set, there's no required base path to talk to the server's controllers.
+     *
+     * @param basePath - The base path to use.
+     *
+     * @returns The builder instance.
+     */
+    base(basePath: string): this {
+      this.#server.#base = basePath;
+
+      return this;
+    }
+
+    /**
+     * Sets the handler that will be called when the server determines
+     * it has no controller that can handle the incoming request.
+     *
+     * If this isn't set, the default handlder returns a 404 response with no body.
+     *
+     * @param handler
+     *
+     * @returns The builder instance.
+     */
+    defaultResponseHandler(handler: (req: Request) => Response): this {
+      this.#server.#defaultResponseHandler = handler;
+
+      return this;
+    }
+
+    /**
+     * Adds the provided middleware to the entry middleware chain.
+     * Entry middleware runs when a request enters the server. It runs before
+     * anything else, and runs even if the client's request didn't include
+     * the server's `base` (if you provided one).
+     *
+     * Middleware runs in the order provided. Async middleware is awaited before
+     * continuing through the chain.
+     *
+     * This method is additive and may be called any number of times. Calling
+     * it more than once won't remove the middleware from previous invocations,
+     * but instead appends the new middleware to the chain.
+     *
+     * @param middleware - The middleware to add.
+     *
+     * @returns The builder instance.
+     */
+    entryMiddleware(...middleware: Middleware[]): this {
+      this.#server.#entryMiddlewareChain.add(...middleware);
+
+      return this;
+    }
+
+    /**
+     * Adds the specified controller to the server. Controllers are the
+     * heart of the server and inform it on what requests it can and cannot
+     * handle.
+     *
+     * If a request comes in and the server can't find an appropriate
+     * request handler in one of its registered controllers, it will send
+     * back the default response.
+     *
+     * @param c - The controller to add.
+     *
+     * @returns The builder instance.
+     */
+    controller(c: Controller): this {
+      this.#server.#controllers.push(c);
+
+      return this;
+    }
+
+    /**
+     * Configures the server to handle HTTPS traffic.
+     *
+     * @param sslOptions - Object containing parameters to customize how the server manages TLS.
+     *
+     * @returns The builder instance.
+     */
+    ssl(sslOptions: Deno.TlsCertifiedKeyPem): this {
+      this.#server.#ssl = sslOptions;
+
+      return this;
+    }
+
+    /**
+     * Subscribes the provided listeners to the server's `onError` event.
+     * Invoking this multiple times will append the new listeners to the
+     * event's subscriptions.
+     *
+     * **Listeners for this event should not throw.**
+     *
+     * @param listeners - The listeners to subscribe.
+     *
+     * @returns The builder instance.
+     */
+    errorListeners(...listeners: ServerErrorListener[]): this {
+      this.#subscribeListeners(this.#server.onError, ...listeners);
+
+      return this;
+    }
+
+    /**
+     * Subscribes the provided listeners to the server's `onDefaultResponse` event.
+     * Invoking this multiple times will append the new listeners to the
+     * event's subscriptions.
+     *
+     * @param listeners - The listeners to subscribe.
+     *
+     * @returns The builder instance.
+     */
+    defaultResponseListeners(...listeners: DefaultResponseListener[]): this {
+      this.#subscribeListeners(this.#server.onDefaultResponse, ...listeners);
+
+      return this;
+    }
+
+    /**
+     * Subscribes the provided listeners to the server's `onBeforeRespond` event.
+     * Invoking this multiple times will append the new listeners to the
+     * event's subscriptions.
+     *
+     * **Listeners for this event should not throw.**
+     *
+     * @param listeners - The listeners to subscribe.
+     *
+     * @returns The builder instance.
+     */
+    beforeRespondListeners(...listeners: BeforeRespondListener[]): this {
+      this.#subscribeListeners(this.#server.onBeforeRespond, ...listeners);
+
+      return this;
+    }
+
+    // deno-lint-ignore no-explicit-any
+    #subscribeListeners<T extends (...args: any[]) => void>(event: Event<T>, ...listeners: T[]): void {
+      listeners.forEach((listener) => event.subscribe(listener));
+    }
+  };
 }