+++ core/lib/test/http-server.test.ts
@@ -1,10 +1,11 @@
 import { assert, assertEquals } from 'assert';
 import { afterAll, afterEach, beforeAll, beforeEach, describe, test } from 'bdd';
 import { type Spy, assertSpyCalls, restore, spy, stub } from 'mock';
+import { Context } from '../context/context.ts';
 import { Controller } from '../controller.ts';
 import { TeapotError } from '../errors/client/teapot.error.ts';
 import { UnauthenticatedError } from '../errors/client/unauthenticated.error.ts';
-import { HttpServer } from '../http-server.ts';
+import { HttpServer, type HttpServerBuilder } from '../http-server.ts';
 import type { Middleware, RequestHandler } from '../model.ts';
 import { JsonResponse } from '../response/json.response.ts';
 
@@ -19,19 +20,25 @@ import { JsonResponse } from '../response/json.response.ts';
  */
 let currentPort = 3000;
 
+const globalContext = new Context(0);
+const testContext = new Context('');
+
 const getBaseRoute = () => `http://localhost:${currentPort}`;
 
 describe('HttpServer', () => {
+  let serverBuilder: HttpServerBuilder;
   let server: HttpServer;
   let entryMiddleware: Spy<Middleware>;
   beforeAll(() => {
     stub(console, 'log');
   });
   beforeEach(async () => {
-    server = new HttpServer();
+    serverBuilder = new HttpServer.Builder();
     entryMiddleware = spy();
 
-    await server.controller(new TestController()).entryMiddleware(entryMiddleware).start(currentPort);
+    server = serverBuilder.controller(new TestController()).entryMiddleware(entryMiddleware).build();
+
+    await server.start(currentPort);
   });
   afterEach(async () => {
     await server.stop();
@@ -59,7 +66,7 @@ describe('HttpServer', () => {
         });
         test(`invokes onDefault when there's no controller to handle the path`, async () => {
           const onDefaultSpy = spy();
-          server.onDefault.subscribe(onDefaultSpy);
+          server.onDefaultResponse.subscribe(onDefaultSpy);
 
           const response = await fetch(`${getBaseRoute()}/nope`);
 
@@ -124,7 +131,7 @@ describe('HttpServer', () => {
 
     describe('with base server path', () => {
       beforeEach(() => {
-        server.base('/api');
+        serverBuilder.base('/api');
       });
 
       describe('defaulting', () => {
@@ -143,7 +150,7 @@ describe('HttpServer', () => {
         });
         test(`invokes onDefault when there's no controller to handle the path`, async () => {
           const onDefaultSpy = spy();
-          server.onDefault.subscribe(onDefaultSpy);
+          server.onDefaultResponse.subscribe(onDefaultSpy);
 
           const response = await fetch(`${getBaseRoute()}/api/nope`);
 
@@ -235,7 +242,7 @@ describe('HttpServer', () => {
     });
     test(`connection information is given to controller middleware`, async () => {
       const controllerMiddleware: Spy<Middleware> = spy();
-      server.controller(new OtherTestController(controllerMiddleware));
+      serverBuilder.controller(new OtherTestController(controllerMiddleware));
 
       const response = await fetch(`${getBaseRoute()}/otherTest`);
 
@@ -258,7 +265,7 @@ describe('HttpServer', () => {
 
   describe('HTTP errors', () => {
     test(`an HTTP error thrown in a global middleware is caught and its status reflected in the response`, async () => {
-      server.entryMiddleware(() => {
+      serverBuilder.entryMiddleware(() => {
         throw new UnauthenticatedError('NONE SHALL PASS!');
       });
 
@@ -269,7 +276,7 @@ describe('HttpServer', () => {
       await cleanupResponses(response);
     });
     test(`an HTTP error thrown in a controller middleware is caught and its status reflected in the response`, async () => {
-      server.controller(
+      serverBuilder.controller(
         new OtherTestController(() => {
           throw new UnauthenticatedError('NONE SHALL PASS!');
         })
@@ -314,7 +321,7 @@ describe('HttpServer', () => {
           }),
         ];
 
-        server.entryMiddleware(...middleware);
+        serverBuilder.entryMiddleware(...middleware);
 
         const response = await fetch(`${getBaseRoute()}/test`);
 
@@ -338,7 +345,7 @@ describe('HttpServer', () => {
           }),
         ];
 
-        server.entryMiddleware(...middleware);
+        serverBuilder.entryMiddleware(...middleware);
 
         const response = await fetch(`${getBaseRoute()}/nope`);
 
@@ -363,7 +370,7 @@ describe('HttpServer', () => {
           }),
         ];
 
-        server.entryMiddleware(...middleware);
+        serverBuilder.entryMiddleware(...middleware);
 
         const response = await fetch(`${getBaseRoute()}/test`);
 
@@ -387,7 +394,7 @@ describe('HttpServer', () => {
           }),
         ];
 
-        server.entryMiddleware(...middleware);
+        serverBuilder.entryMiddleware(...middleware);
 
         const response = await fetch(`${getBaseRoute()}/test`);
 
@@ -411,7 +418,7 @@ describe('HttpServer', () => {
           }),
         ];
 
-        server.controller(new OtherTestController(...middleware));
+        serverBuilder.controller(new OtherTestController(...middleware));
 
         const response = await fetch(`${getBaseRoute()}/otherTest`);
 
@@ -447,8 +454,8 @@ describe('HttpServer', () => {
           }),
         ];
 
-        server.controller(new OtherTestController(...controllerMiddleware));
-        server.entryMiddleware(...entryMiddleware);
+        serverBuilder.controller(new OtherTestController(...controllerMiddleware));
+        serverBuilder.entryMiddleware(...entryMiddleware);
 
         const response = await fetch(`${getBaseRoute()}/nope`);
 
@@ -470,7 +477,7 @@ describe('HttpServer', () => {
           spy(() => {}),
         ];
 
-        server.controller(new OtherTestController(...middleware));
+        serverBuilder.controller(new OtherTestController(...middleware));
 
         const response = await fetch(`${getBaseRoute()}/otherTest`);
 
@@ -493,7 +500,7 @@ describe('HttpServer', () => {
           }),
         ];
 
-        server.controller(new OtherTestController(...middleware));
+        serverBuilder.controller(new OtherTestController(...middleware));
 
         const response = await fetch(`${getBaseRoute()}/otherTest`);
 
@@ -508,7 +515,7 @@ describe('HttpServer', () => {
 
     describe('headers', () => {
       test(`headers added in global middleware appear in the response`, async () => {
-        server.entryMiddleware(({ resHeaders }) => resHeaders.set('custom', 'jt'));
+        serverBuilder.entryMiddleware(({ resHeaders }) => resHeaders.set('custom', 'jt'));
 
         const response = await fetch(`${getBaseRoute()}/test`);
 
@@ -517,7 +524,7 @@ describe('HttpServer', () => {
         await cleanupResponses(response);
       });
       test(`headers added in controller middleware appear in the response`, async () => {
-        server.controller(new OtherTestController(({ resHeaders }) => resHeaders.set('custom', 'jt')));
+        serverBuilder.controller(new OtherTestController(({ resHeaders }) => resHeaders.set('custom', 'jt')));
 
         const response = await fetch(`${getBaseRoute()}/otherTest`);
 
@@ -526,8 +533,8 @@ describe('HttpServer', () => {
         await cleanupResponses(response);
       });
       test(`a header set in controller middleware overrides the same one set in global middleware`, async () => {
-        server.controller(new OtherTestController(({ resHeaders }) => resHeaders.set('custom', 'jt')));
-        server.entryMiddleware(({ resHeaders }) => resHeaders.set('custom', 'global jt'));
+        serverBuilder.controller(new OtherTestController(({ resHeaders }) => resHeaders.set('custom', 'jt')));
+        serverBuilder.entryMiddleware(({ resHeaders }) => resHeaders.set('custom', 'global jt'));
 
         const response = await fetch(`${getBaseRoute()}/otherTest`);
 
@@ -536,8 +543,10 @@ describe('HttpServer', () => {
         await cleanupResponses(response);
       });
       test(`a header set in the response of the controller's handler overrides the same one set in other middleware`, async () => {
-        server.controller(new OtherTestController(({ resHeaders }) => resHeaders.set('server-custom-header', 'jt')));
-        server.entryMiddleware(({ resHeaders }) => resHeaders.set('server-custom-header', 'global jt'));
+        serverBuilder.controller(
+          new OtherTestController(({ resHeaders }) => resHeaders.set('server-custom-header', 'jt'))
+        );
+        serverBuilder.entryMiddleware(({ resHeaders }) => resHeaders.set('server-custom-header', 'global jt'));
 
         const response = await fetch(`${getBaseRoute()}/otherTest`);
 
@@ -546,8 +555,8 @@ describe('HttpServer', () => {
         await cleanupResponses(response);
       });
       test(`a header set in the response of the controller is present alongside those set in middleware`, async () => {
-        server.controller(new OtherTestController(({ resHeaders }) => resHeaders.set('custom-header', 'jt')));
-        server.entryMiddleware(({ resHeaders }) => resHeaders.set('another-custom-header', 'global jt'));
+        serverBuilder.controller(new OtherTestController(({ resHeaders }) => resHeaders.set('custom-header', 'jt')));
+        serverBuilder.entryMiddleware(({ resHeaders }) => resHeaders.set('another-custom-header', 'global jt'));
 
         const response = await fetch(`${getBaseRoute()}/otherTest`);
 
@@ -558,7 +567,7 @@ describe('HttpServer', () => {
         await cleanupResponses(response);
       });
       test(`a header set in the repsonse of a middleware is present alongside those set in other middleware`, async () => {
-        server.entryMiddleware(
+        serverBuilder.entryMiddleware(
           ({ resHeaders }) => resHeaders.set('custom-header', 'jt'),
           () => new Response(undefined, { headers: { 'other-header': 'response-header' } })
         );
@@ -575,7 +584,7 @@ describe('HttpServer', () => {
 
   describe('headers', () => {
     test(`custom headers are present`, async () => {
-      server.controller(new OtherTestController());
+      serverBuilder.controller(new OtherTestController());
 
       const response = await fetch(`${getBaseRoute()}/otherTest`);
 
@@ -635,13 +644,118 @@ describe('HttpServer', () => {
   });
 
   test(`uses default response handler when provided`, async () => {
-    server.defaultResponseHandler(() => new Response(undefined, { status: 401 }));
+    serverBuilder.defaultResponseHandler(() => new Response(undefined, { status: 401 }));
 
     const response = await fetch(`${getBaseRoute()}/nope`);
 
     assertEquals(response.status, 401);
 
     await cleanupResponses(response);
+  });
+
+  describe('context', () => {
+    beforeEach(() => {
+      serverBuilder.entryMiddleware(({ setContext }) => {
+        setContext(globalContext, 123);
+      });
+    });
+
+    describe('can read global context...', () => {
+      test(`inside entry middleware`, async () => {
+        serverBuilder.entryMiddleware(({ getContext }) => {
+          return new JsonResponse({ ctx: getContext(globalContext) });
+        });
+
+        const response = await fetch(`${getBaseRoute()}/test`);
+
+        const json = await response.json();
+
+        assertEquals(response.status, 200);
+        assertEquals(json.ctx, 123);
+
+        await cleanupResponses(response);
+      });
+      test(`inside controller middleware`, async () => {
+        serverBuilder.controller(
+          new ControllerWithContext(({ getContext }) => {
+            return new JsonResponse({ ctx: getContext(testContext) });
+          })
+        );
+
+        const response = await fetch(`${getBaseRoute()}/contextTest`);
+
+        const json = await response.json();
+
+        assertEquals(response.status, 200);
+        assertEquals(json.ctx, '321');
+
+        await cleanupResponses(response);
+      });
+      test(`inside request handler`, async () => {
+        serverBuilder.controller(new ControllerWithContext());
+
+        const response = await fetch(`${getBaseRoute()}/contextTest/global`);
+
+        const json = await response.json();
+
+        assertEquals(response.status, 200);
+        assertEquals(json.ctx, 123);
+
+        await cleanupResponses(response);
+      });
+    });
+
+    describe('writes work...', () => {
+      test(`from inside entry middleware middleware`, async () => {
+        serverBuilder.entryMiddleware(
+          ({ setContext }) => {
+            setContext(globalContext, 1234);
+          },
+          ({ getContext }) => {
+            return new JsonResponse({
+              ctx: getContext(globalContext),
+            });
+          }
+        );
+
+        const response = await fetch(`${getBaseRoute()}/test`);
+
+        const json = await response.json();
+
+        assertEquals(response.status, 200);
+        assertEquals(json.ctx, 1234);
+
+        await cleanupResponses(response);
+      });
+      test(`from inside controller middleware`, async () => {
+        serverBuilder.controller(
+          new ControllerWithContext(({ setContext }) => {
+            setContext(globalContext, 321);
+          })
+        );
+
+        const response = await fetch(`${getBaseRoute()}/contextTest/global`);
+
+        const json = await response.json();
+
+        assertEquals(response.status, 200);
+        assertEquals(json.ctx, 321);
+
+        await cleanupResponses(response);
+      });
+      test(`from inside controller middleware when the context doesn't exist globally`, async () => {
+        serverBuilder.controller(new ControllerWithContext());
+
+        const response = await fetch(`${getBaseRoute()}/contextTest`);
+
+        const json = await response.json();
+
+        assertEquals(response.status, 200);
+        assertEquals(json.ctx, '321');
+
+        await cleanupResponses(response);
+      });
+    });
   });
 });
 
@@ -695,6 +809,28 @@ class TestController extends Controller {
 
   'GET /checkRemoteAddr': RequestHandler = ({ remoteAddr }) => {
     return new JsonResponse({ present: !!remoteAddr });
+  };
+}
+
+class ControllerWithContext extends Controller {
+  constructor(...middleware: Middleware[]) {
+    super({
+      base: '/contextTest',
+      middleware: [
+        ({ setContext }) => {
+          setContext(testContext, '321');
+        },
+        ...middleware,
+      ],
+    });
+  }
+
+  'GET /': RequestHandler = ({ getContext }) => {
+    return new JsonResponse({ ctx: getContext(testContext) });
+  };
+
+  'GET /global': RequestHandler = ({ getContext }) => {
+    return new JsonResponse({ ctx: getContext(globalContext) });
   };
 }