+++ logging/lib/mod.ts
@@ -1,3 +1,3 @@
+export * from './logger.context.ts';
 export * from './logger.ts';
 export * from './logging.middleware.ts';
-export * from './logging.model.ts';