+++ core/changelog.md
@@ -1,3 +1,47 @@
+# 0.7.0
+
+## Breaking Changes
+
+- The type `ServerErrorHandler` has been renamed to `ServerErrorListener`.
+- The type `DefaultResponseHandler` has been renamed to `DefaultResponseListener`.
+- The type `BeforeRespondHandler` has been renamed to `BeforeRespondListener`.
+- The `onDefault` event has been renamed to `onDefaultResponse`.
+- The `remoteAddr` given to middleware and `RequestHandler`s is now of type `Deno.NetAddr` instead of `Deno.Addr`.
+- Creating a server has been tweaked
+  - Instead of invoking `new HttpServer()` and calling chaining building methods on it (`base`, `controller`, etc.), that building logic has been moved to a dedicated builder to clean up and separate the `HttpServer`'s methods from its building methods. The required change is small and looks like this:
+      ```ts
+      // <0.7.0
+      new HttpServer()
+        .base(...)
+        .entryMiddlware(...)
+        .controller(...)
+        .start(3000);
+
+      // 0.7.0
+      new HttpServer.Builder() // Now creating an instance of `HttpServer.Builder`
+        .base(...)
+        .entryMiddlware(...)
+        .controller(...)
+        .build() // `build`ing returns the underlying server the builder has been making.
+        .start(3000);
+      ```
+  The new `HttpServer.Builder` has had all the server-building methods moved to it, so your bootstrapping code should only have to change very slightly. 
+    - Additionally, **3 new building methods** have been added: `errorListeners`, `defaultResponseListeners`, and `beforeRespondListeners`. These methods will subscribe listeners to the server's `onError`, `onDefaultResponse`, and `onBeforeRespond` events, respectively.
+      - You may still interact directly with the server's exposed instances of these events if you prefer.
+  - For your convenience, if you need to annotate the type of an `HttpServer.Builder` instance, there is now an `HttpServerBuilder` type exported from the lib.
+
+- Context has been reworked
+  - The previous implementation created a type nightmare and having a highly-volatile super object that served as a dumping ground for any and all app information associated to a request irked me. It also added boilerplate code to guarantee strict typing that was cumbersome, and ensuring this magical super object was correctly typed was annoying at best.
+  - `Middleware` and `RequestHandler` have been updated.
+    - `Middleware` now receives two new functions: `getContext` and `setContext`.
+    - `RequestHandler` now receives one new function: `getContext`
+  - To update your application:
+    - You can remove any custom context types, and the structures that used to accept a generic argument for the sake of context no longer take a generic argument.
+      - You may now yeet any custom type you had for passing to these generic arguments.
+    - Instead of setting context by setting a propery on `ctx`, you now create a context with `new Context`. You'll use that instance for all the new operations with context, including setting it. Use the `setContext` function now available to middleware with your `Context` instance to set the value in that scope.
+    - Instead of reading properties off `ctx`, you now use `getContext` in conjunction with a `Context` instance. You'll receive the value for the current scope.
+  - **For more information on what context is, what problem it solves, and how to use it, refer to the new `Context` section in the README.**
+
 # 0.6.0
 
 ## New Features